requirements.txt
django==3.2
djangorestframework
mysqlclient
django-cors-headers
python-decouple
gunicorn
whitenoise